+++ packages/collapsible-panel/src/routes/+page.svelte
@@ -12,7 +12,3 @@
 	<p>Create your package using @sveltejs/package and preview/showcase your work with SvelteKit</p>
 	<p>Visit <a href="https://kit.svelte.dev">kit.svelte.dev</a> to read the documentation</p>
 </CollapsiblePanel>
-
-<style>
-	@import 'https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma.min.css';
-</style>