+++ packages/style-switcher/src/routes/+page.svelte
@@ -7,7 +7,3 @@
 </sveltekit:head>
 
 <Map />
-
-<style>
-	@import 'https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma.min.css';
-</style>