+++ packages/legend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @watergis/svelte-maplibre-legend
 
+## 0.1.6
+
+### Patch Changes
+
+- 74d94ad: deleted bulma dependency from legend control
+
+## 0.1.5
+
+### Patch Changes
+
+- 47bb80f: Use bulma.sass instead of CDN
+
 ## 0.1.4
 
 ### Patch Changes